+++ odbc/src/api/col_attr_describe_tests.rs
@@ -171,7 +171,7 @@ mod unit {
             let mut data_type = SqlDataType::SQL_UNKNOWN_TYPE;
             let col_size = &mut 42usize;
             let decimal_digits = &mut 42i16;
-            let mut nullable = Nullability::SQL_NO_NULLS;
+            let mut nullable = Nullability::SQL_NO_NULLS as i16;
             // test string attributes
             assert_eq!(
                 SqlReturn::ERROR,
@@ -301,7 +301,7 @@ mod unit {
                 let mut data_type = SqlDataType::SQL_UNKNOWN_TYPE;
                 let col_size = &mut 42usize;
                 let decimal_digits = &mut 42i16;
-                let mut nullable = Nullability::SQL_NO_NULLS;
+                let mut nullable = Nullability::SQL_NO_NULLS as i16;
                 // test string attributes
                 assert_eq!(
                     SqlReturn::ERROR,
@@ -599,7 +599,7 @@ mod unit {
             let mut data_type = SqlDataType::SQL_UNKNOWN_TYPE;
             let col_size = &mut 42usize;
             let decimal_digits = &mut 42i16;
-            let mut nullable = Nullability::SQL_NULLABLE_UNKNOWN;
+            let mut nullable = Nullability::SQL_NULLABLE_UNKNOWN as i16;
             // test string attributes
             assert_eq!(
                 SqlReturn::SUCCESS,
@@ -624,7 +624,7 @@ mod unit {
             // decimal_digits should be 0
             assert_eq!(0i16, *decimal_digits);
             // nullable should stay as NO_NULLS
-            assert_eq!(Nullability::SQL_NO_NULLS, nullable);
+            assert_eq!(Nullability::SQL_NO_NULLS as i16, nullable);
             // name_buffer should contain TABLE_NAME
             assert_eq!(
                 "TABLE_NAME".to_string(),
